macos: widen sub_GAME_7F08BEEC's matrix stride (fixes Dam/Facility/Surface)
`(Mtxf *)((u32)matrices + j)` truncated the prop-matrix base before the
per-instance multiply; re-based to uintptr_t. Dam, Facility, Surface1 and
Surface2 all now run their full window. Three signatures left: Statue/Jungle
(modelConstrainOrWrapAnimFrame), Archives/Control (strncpy), Cuba
(bondviewLoadSetupIntroSection).

+++ src/game/bondview2.c
@@ -10883,7 +10883,7 @@ void sub_GAME_7F08BEEC(Mtxf *matrices, s32 count)
 
     for (i = 0, j = 0; i < count; i++, j += sizeof(Mtxf))
     {
-        matrix_4x4_multiply_homogeneous(currentPlayerGetViewToWorldMtxf(), (Mtxf *)((u32)matrices + j), &sp40);
+        matrix_4x4_multiply_homogeneous(currentPlayerGetViewToWorldMtxf(), (Mtxf *)((uintptr_t)matrices + j), &sp40); /* D298/M2 */
 
         sp40.m[3][0] -= g_CurrentPlayer->current_model_pos.f[0];
         sp40.m[3][1] -= g_CurrentPlayer->current_model_pos.f[1];
